bump rake and enable running Ruby in `-w` mode during tests
Also fix the plethora of warnings that are emitted. The reason for this is
to be able to run Ruby in `-w` mode in the test suites of apps that consume
this gem without stdout being flooded with irrelevant noise. For an example
of this, see recent test runs in zendesk_proxy. The sources of warnings were
mostly access to uninitialised ivars, which is fixed by defined? checks and
attr_accessors, and a misconfiguration that caused Rails 5.1 to try to access
the entire filesystem, generating enough permissions warnings to crash Travis.