Introduce ActiveRecordShards::ShardedModel
From now on sharded models must inherit from ActiveRecordShards::ShardedModel
- and no longer will users need to use the "not_sharded" instruction to mark
models not sharded.

By doing this we've simplified the logic around sharding as
ActiveRecord::Base no longer needs to be part of the calculation around
wether a class is sharded or not. It's now purely defined in terms of
inheritance, inheriting from ShardedModel means the model is sharded.