Simplify is_sharded? calculation further
Instead of checking class inheritance we can just call a class method. This way
we can define it to be false for ActiveRecord::Base and override it in sharded
subclasses. This also gives us the possibility of using a module (or just plain
defining the method in any random class).