Use logging for deprecation
ActiveSupport::Deprecation may potentially raise exception in applications
using this gem if Rails is configured to do so. This is very hard for users of
the gem to do anything about apart from not raising on deprecations. To avoid
this we'll revert to just logging deprecation notices for now. Once all
libraries using the not_sharded method have been upgraded we can fully
deprecate the method.