Handle the case of switching to no shard
When returning from a scope where a shard is selected to a scope where no shard
is selected we want to return the system to that same state. This ensures that
sharded database calls are always performed inside scopes where a shard is
selected and thus eliminates the risk that a user may accidentally try to
perform database queries against an uninitialized shard selection.