Use a fresh connection handler before tests
The connection handler is normally a singleton on `ActiveRecord::Base`, and so
would cache `@owner_to_pool` between tests. And since we run tests with
different database configurations etc., we should start out each test run with
a fresh connection handler.

Second, in the `ActiveRecordShards::ConnectionSwitcher` module that we define
when running with Rails 4 and below, we cache the database specifications in a
class variable `@@specification_cache` which would be re-used between test
runs. This cache must be cleared between tests, especially those that use
different database configurations.