Use a TCP Proxy to disable unsharded primary
When we need to test that a query is performed against a replica and not
the primary, we have traditionally done so by disconnecting, changing
the DB config hash, and reconnection - and doing it in reverse after
the test.

With this change, we will instead connect to the unsharded primary DB
via a proxy, which can then be disabled and re-enabled before and after
such tests. This gives us (or at least me) much more confidence that our
tests will keep working across all Rails versions.