Don't let disabled TCPProxy hang indefinitely
Test-only related change: If the TCPProxy receives a request while
paused, we would previously do nothing - just wait 0.2 seconds and then
read some more from the socket, ad inifitum. That meant that if our code
under test would query the wrong database (through a paused proxy), the
specs would hand indefinitely.

With this commit, the implementation is changed for the "requests"
thread, so that if our tests makes a query (via `Mysql2::Client` or the
`mysql` shell command etc.), an exception is raised and the test will
fail immediately.