Fix replica schema loading for Rails 6.0
In Rails 6.0, when loading an association for which AR doesn't yet know
the full schema, we would see the schema potentially being loaded from
the primary database with a stacktrace including:

  AR/connection_adapters/abstract/schema_statements.rb:37:in `data_sources'
  AR/connection_adapters/schema_cache.rb:137:in `prepare_data_sources'
  AR/connection_adapters/schema_cache.rb:54:in `data_source_exists?'
  AR/type_caster/connection.rb:21:in `type_for_attribute'
  AR/table_metadata.rb:33:in `type'
  AR/relation/predicate_builder.rb:51:in `build'

That is, `SchemaCache#data_source_exists?` is being called from
somewhere that wasn't yet patched. The code was introduced in
https://github.com/rails/rails/pull/36314

This here commit changes how `ActiveRecord::TypeCaster::Connection`
resolves a `connection`, which usually is just delegated to `@klass`.
Here, we implement a patch to mimic the
`ActiveRecordShards::DefaultReplicaPatches#on_replica_unless_tx` method
which unfortunately expects a block argument.

---

Unfortunately, Rails 4 is still allowed to read the schema from the
primary database.