Disconnect from DB before pausing TCP proxy
When testing e.g. the DefaultReplicaPatches, we pause the TCP proxy,
ensuring that no requests are sent to MySQL. But the ActiveRecord
connection might still be used, e.g. for escaping SQL strings and the
like. When such a test was the first test to be executed, ActiveRecord
would try to make an actual connection to MySQL, which would be caught
by the paused proxy.

We don't like flaky tests, so let's disconnect all existing connections
before we pause the proxy, making sure we discover all potential
attempts at opening new connections.