Memoize Active Record configurations
Rails 6 changed ActiveRecord configurations from a hash to a custom
object. You can still access the configurations with Hash syntax but it
gives deprecation warnings and is a lot slower.

This PR memoizes the configurations so we only have to do the slow
access once. Needs to be changed to the new syntax (configs_for…) before
that API is removed.

We updated the performance test with new results showing Rails 6.0 being
significantly faster.

Co-authored-by: Nony Dutton <nony.dutton@zendesk.com>