Remove with_default_shard
The patches for `load_schema!` and `table_exists?` were made to ensure
that sharded models could still call these methods even if no connection
was currently selected. That should work out of the box now.

Or does it? We connect to `:default` by default, and that isn’t a
sharded database.