Bump version to v5.0.0
Yes, we didn't have any v4.0.0 release, but we did have a lot of
v4.0.0-beta* releases a few years ago. And we don't want eventual users
of one of those releases to think _this_ is the release they have been
waiting for, finally out of beta.

No, this release is a breaking change both for users on 3.x.x and users
on 4.0.0-beta*.