Remove Mocha gem
Mocha is a large gem, and it includes e.g. `any_instance` which seems like an
anti-pattern. At least I think that when `any_instance` is in use, I don't
_really_ know what is going on during the test. If you can define and return
(and stub on) _one_ instance, things are much easier to follow.

I was about to add the "minitest-mock_expectations" gem to replace Mocha's
`#expects` method when I realised that in those cases we actually only need the
stub - no mocking is necessary.