Fix flaky tests introduced by caching is_sharded?
The issue is that some tests are resetting RAILS_ENV, which resets the
database configuration which might actually change the is sharded status
of the models.

Introduce a reset_app_env! method to bust these caches; mostly for
tests, but if someone is doing something nasty, this will fix it for
them.