Fix SqlComments
`SqlComments` is broken in Rails 7.0 because it now passes an `async`
kwarg[^1] when it used to only have the two positional arguments.

[^1]:(https://github.com/rails/rails/blob/5873d6279a2b2ec11d0176da66df17b076733791/activerecord/lib/active_record/connection_adapters/abstract_mysql_adapter.rb#L194)

The reason I'm using a sublass of `Mysql2Adapter` is so that we can
prepend `SqlComments` onto it and use the instance of that custom class
to test that the comments are added without adding comments to the rest
of the test suite.

I've updated the tests to... actually test with Rails. Before it didn't
run any Rails code at all.